src/backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.17
Werkzeug==1.0.1
mss==2.0.22
numpy==1.18.4
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.1.2
